requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
PyYAML>=6.0
rumps>=0.4.0
